refactor(reports): add new reports folder, create timeOptions and transactionList component, separate responsabilities
